Remove jsonpatch dependency
The recently discovered bug with json patch merge type was initially
addressed by applying the json patch client-side. This change instead
ensures that the json patch is passed through to the kubernetes api
unmodified. As this addresses the root cause of the bug, the jsonpatch
dependency is no longer necessary.