helm: List all matching releases, regardless of state
There are cases where a `helm list` does not list all
existing releases. One such case is when a release
is in state pending-upgrade, or any other pending state.

Now with this change any use of the lookup or info modules
will now return such releases, so that one can work with them.