Fix sanity tests
This rewrites the sanity ignores refresh script to fully manage the
ignores files. Previously, the ignore files were being handled by a mix
of manual and automated. This was tedious and error prone. This new
refresh script will overwrite existing ignore files, so changes to the
ignore files should all be done in the script moving forward.