Improved refactor of hide_fields to better cope
Needed to cope with multiple levels of bracketed keys

e.g. spec.template.spec[securityContext][runAsUser]

and implementing that lead to a much simpler version
that finds the first key and the rest, and deals with
just those two.

Added more tests to cover some of the new supported
cases.