Implement integration test for `plain_http` param
In order to test an insecure http registry, one needs to be accessible
from the test environment. `kind-action` provides a simple way to run
such a registry, so an helm chart needs to get pushed to it. After that
it's just as simple as trying to install the chart from that local
registry

Signed-off-by: Matteo Danelon <matteo.danelon@aleasrl.com>