add validate-modules:bad-return-value-key to ignore list
Return value key 'values', 'items', and 'keys' are not recomended
for module return keys and starting from 2.21 they will raise a warning
in the ansible-test, ref. https://github.com/ansible/ansible/pull/86079

Proper way is to rename it the modules:
- plugins/modules/helm.py
- plugins/modules/helm_info.py
- plugins/modules/k8s.py

However, to keep backward compatibility we need to keep the them in the
current `stable-6` and `stable-5` branches and this commit is to
stabilize CI for the time being.

Resolve: #1050